+++ src/routes/comment.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import { PrismaClient } from '@prisma/client';
 import jwtMiddleware from '../middlewares/jwtMiddleware.js';
-import validateCommentMiddleware from '../middlewares/validateCommentMiddleware.js';
+import validateCommentMiddleware from '../middlewares/validate/validateCommentMiddleware.js';
 import * as commentController from '../controllers/commentController.js';
 
 const router = express.Router();
