+++ src/routes/user.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import { PrismaClient } from '@prisma/client';
 import * as userController from '../controllers/userController.js';
-import * as validateUserMiddleware from '../middlewares/validateUserMiddleware.js';
+import * as validateUserMiddleware from '../middlewares/validate/validateUserMiddleware.js';
 import jwtMiddleware from '../middlewares/jwtMiddleware.js';
 import cookieParser from 'cookie-parser';
